requirements.txt
torch>=2.0
torch-geometric>=2.4
tweepy>=4.14
sentence-transformers>=2.2
scikit-learn>=1.3
networkx>=3.0
matplotlib>=3.7
seaborn>=0.12
tqdm
python-dotenv
